+++ showcase/proto-gapic-showcase-v1beta1/src/main/java/com/google/showcase/v1beta1/StreamingSequenceReportName.java
@@ -129,7 +129,7 @@ public class StreamingSequenceReportName implements ResourceName {
     if (o == this) {
       return true;
     }
-    if (o != null || getClass() == o.getClass()) {
+    if (o != null && getClass() == o.getClass()) {
       StreamingSequenceReportName that = ((StreamingSequenceReportName) o);
       return Objects.equals(this.streamingSequence, that.streamingSequence);
     }